requirements.txt
graphviz
pillow